feat: return back first solution 01-simple-tests; rename files; add variants
