+++ package.json
@@ -13,6 +13,7 @@
     "test": "jest --silent",
     "test:verbose": "jest",
     "test:01-simple-tests": "jest src/01-simple-tests/index.test.ts",
+    "test:01-simple-tests-2": "jest src/01-simple-tests/index.spec.ts",
     "test:02-table-tests": "jest src/02-table-tests/index.test.ts",
     "test:03-error-handling-async": "jest src/03-error-handling-async/index.test.ts",
     "test:04-test-class": "jest src/04-test-class/index.test.ts",
