+++ src/01-simple-tests/index.test.ts
@@ -1,5 +1,7 @@
-import { random } from 'lodash';
 import { simpleCalculator, Action } from './index';
+
+// VARIANT 1 (simple)
+// ---------------------------------
 
 // It's a pity that this type cannot be imported from the 'index.ts' file.
 type RawCalculatorInput = {
@@ -8,76 +10,73 @@ type RawCalculatorInput = {
   action: unknown;
 };
 
-const getRandomNumber = (min: number, max: number): number => {
-  return random(min, max, false);
-};
-
-const getRandomInvalid = (): unknown => {
-  const invalidValues = [
-    'string',
-    null,
-    undefined,
-    true,
-    false,
-    [],
-    {},
-    () => NaN,
-    Symbol('sym'),
-    // NaN,
-  ];
-  return invalidValues[Math.floor(Math.random() * invalidValues.length)];
-};
-
-const getRandomInput = (
-  action: unknown,
-  invalidArg = false,
-): RawCalculatorInput => {
-  return {
-    a: invalidArg ? getRandomInvalid() : getRandomNumber(-100, 100),
-    b: invalidArg ? getRandomInvalid() : getRandomNumber(-100, 100),
-    action: action,
-  };
-};
-
 describe('simpleCalculator tests', () => {
   test('should add two numbers', () => {
-    const input = getRandomInput(Action.Add);
+    const input: RawCalculatorInput = {
+      a: 1,
+      b: 2,
+      action: Action.Add,
+    };
     const result = simpleCalculator(input);
-    expect(result).toBe(Number(input.a) + Number(input.b));
+    expect(result).toBe(3);
   });
 
   test('should subtract two numbers', () => {
-    const input = getRandomInput(Action.Subtract);
+    const input: RawCalculatorInput = {
+      a: 9,
+      b: 5,
+      action: Action.Subtract,
+    };
     const result = simpleCalculator(input);
-    expect(result).toBe(Number(input.a) - Number(input.b));
+    expect(result).toBe(4);
   });
 
   test('should multiply two numbers', () => {
-    const input = getRandomInput(Action.Multiply);
+    const input: RawCalculatorInput = {
+      a: 2,
+      b: 2,
+      action: Action.Multiply,
+    };
     const result = simpleCalculator(input);
-    expect(result).toBe(Number(input.a) * Number(input.b));
+    expect(result).toBe(4);
   });
 
   test('should divide two numbers', () => {
-    const input = getRandomInput(Action.Divide);
+    const input: RawCalculatorInput = {
+      a: 9,
+      b: 3,
+      action: Action.Divide,
+    };
     const result = simpleCalculator(input);
-    expect(result).toBe(Number(input.a) / Number(input.b));
+    expect(result).toBe(3);
   });
 
   test('should exponentiate two numbers', () => {
-    const input = getRandomInput(Action.Exponentiate);
+    const input: RawCalculatorInput = {
+      a: 2,
+      b: 3,
+      action: Action.Exponentiate,
+    };
     const result = simpleCalculator(input);
-    expect(result).toBe(Number(input.a) ** Number(input.b));
+    expect(result).toBe(8);
   });
 
   test('should return null for invalid action', () => {
-    const input = getRandomInput(getRandomInvalid());
+    const input: RawCalculatorInput = {
+      a: 1,
+      b: 3,
+      action: 'invalid',
+    };
     const result = simpleCalculator(input);
     expect(result).toBeNull();
   });
 
   test('should return null for invalid arguments', () => {
-    const input = getRandomInput(Action.Add, true);
+    const input: RawCalculatorInput = {
+      a: 'invalid',
+      b: 1,
+      action: Action.Add,
+    };
     const result = simpleCalculator(input);
     expect(result).toBeNull();
   });
