+++ src/07-mocking-lib-api/index.test.ts
@@ -1,10 +1,13 @@
 import axios from 'axios';
-import { throttledGetDataFromApi, THROTTLE_TIME } from './index';
+import { throttledGetDataFromApi } from './index';
 
-// VARIANT 1 (with `useFakeTimers`)
+// VARIANT 1 (with mock `throttle`)
 // ---------------------------------
 
 jest.mock('axios');
+jest.mock('lodash', () => ({
+  throttle: jest.fn((fn) => fn),
+}));
 
 describe('throttledGetDataFromApi', () => {
   let mockAxiosInstance: jest.Mocked<typeof axios>;
@@ -13,11 +16,6 @@ describe('throttledGetDataFromApi', () => {
     mockAxiosInstance = axios as jest.Mocked<typeof axios>;
     mockAxiosInstance.create.mockReturnThis();
     mockAxiosInstance.get.mockResolvedValue({ data: 'mock data' });
-    jest.useFakeTimers();
-  });
-
-  afterEach(() => {
-    jest.useRealTimers();
   });
 
   test('should create instance with provided base url', async () => {
@@ -29,35 +27,12 @@ describe('throttledGetDataFromApi', () => {
 
   test('should perform request to correct provided url', async () => {
     const relativePath = '/test';
-    jest.advanceTimersByTime(THROTTLE_TIME);
     await throttledGetDataFromApi(relativePath);
     expect(mockAxiosInstance.get).toHaveBeenCalledWith(relativePath);
   });
 
   test('should return response data', async () => {
-    jest.advanceTimersByTime(THROTTLE_TIME);
     const data = await throttledGetDataFromApi('/test');
     expect(data).toBe('mock data');
   });
-
-  // THROTTLE TEST:
-  // ---------------------------------------------------------------------------
-  /*   
-    throttle options -> {
-      leading: true,  // call immediately
-      trailing: false // Do not call after throttling ends
-    }); 
-  */
-  /*   
-    test('should throttle function calls', async () => {
-    const relativePath = '/test';
-    await throttledGetDataFromApi(relativePath);
-    expect(mockAxiosInstance.get).toHaveBeenCalledWith(relativePath);
-    await throttledGetDataFromApi(relativePath);
-    expect(mockAxiosInstance.get).toHaveBeenCalledTimes(1);
-    jest.advanceTimersByTime(THROTTLE_TIME);
-    await throttledGetDataFromApi(relativePath);
-    expect(mockAxiosInstance.get).toHaveBeenCalledTimes(2 + 1); // + 1 call after throttling ends
-  }); 
-  */
 });
